+++ disasm_scalar32.cpp
@@ -16,10 +16,10 @@ namespace disasm {
                     do {
                         rv = new scalar32_insn(string("addcmpb")+cc,
                                                "r{d}, r{a}, r{s}, $+({o}*2)");
-                        vc4_parameter d0(ParameterTypes::REGISTER, (insn >> 16) & 0x0000000f);
-                        vc4_parameter a0(ParameterTypes::REGISTER, (insn >> 16) & 0x000000f0);
-                        vc4_parameter s0(ParameterTypes::REGISTER, (insn >> 10) & 0x0000000f);
-                        vc4_parameter o0(ParameterTypes::IMMEDIATE, insn & 0x000003ff);
+                        vc4_parameter d0(ParameterTypes::REGISTER, (uint32_t)((insn >> 16) & 0x0000000f));
+                        vc4_parameter a0(ParameterTypes::REGISTER, (uint32_t)((insn >> 16) & 0x000000f0));
+                        vc4_parameter s0(ParameterTypes::REGISTER, (uint32_t)((insn >> 10) & 0x0000000f));
+                        vc4_parameter o0(ParameterTypes::IMMEDIATE, (uint32_t)(insn & 0x000003ff));
                         rv->addParameter("d", d0)->addParameter("a", a0)
                             ->addParameter("s", s0)->addParameter("o", o0);
                     } while(0);
@@ -28,10 +28,10 @@ namespace disasm {
                     do {
                         rv = new scalar32_insn(string("addcmpb")+cc,
                                                "r{d}, {i}, r{s}, $+({o}*2)");
-                        vc4_parameter d1(ParameterTypes::REGISTER, (insn >> 16) & 0x0000000f);
-                        vc4_parameter i0(ParameterTypes::REGISTER, (insn >> 16) & 0x000000f0);
-                        vc4_parameter s1(ParameterTypes::REGISTER, (insn >> 10) & 0x0000000f);
-                        vc4_parameter o1(ParameterTypes::IMMEDIATE, insn & 0x000003ff);
+                        vc4_parameter d1(ParameterTypes::REGISTER, (uint32_t)((insn >> 16) & 0x0000000f));
+                        vc4_parameter i0(ParameterTypes::REGISTER, (uint32_t)((insn >> 16) & 0x000000f0));
+                        vc4_parameter s1(ParameterTypes::REGISTER, (uint32_t)((insn >> 10) & 0x0000000f));
+                        vc4_parameter o1(ParameterTypes::IMMEDIATE, (uint32_t)(insn & 0x000003ff));
                         rv->addParameter("d", d1)->addParameter("i", i0)
                             ->addParameter("s", s1)->addParameter("o", o1);
                     } while(0);
@@ -40,10 +40,10 @@ namespace disasm {
                     do {
                         rv = new scalar32_insn(string("addcmpb")+cc,
                                                "r{d}, r{a}, {u}, $+({o}*2)");
-                        vc4_parameter d2(ParameterTypes::REGISTER, (insn >> 16) & 0x0000000f);
-                        vc4_parameter a1(ParameterTypes::REGISTER, (insn >> 16) & 0x000000f0);
-                        vc4_parameter u0(ParameterTypes::REGISTER, (insn >> 8) & 0x0000003f);
-                        vc4_parameter o2(ParameterTypes::IMMEDIATE, insn & 0x000000ff);
+                        vc4_parameter d2(ParameterTypes::REGISTER, (uint32_t)((insn >> 16) & 0x0000000f));
+                        vc4_parameter a1(ParameterTypes::REGISTER, (uint32_t)((insn >> 16) & 0x000000f0));
+                        vc4_parameter u0(ParameterTypes::REGISTER, (uint32_t)((insn >> 8) & 0x0000003f));
+                        vc4_parameter o2(ParameterTypes::IMMEDIATE, (uint32_t)(insn & 0x000000ff));
                         rv->addParameter("d", d2)->addParameter("a", a1)
                             ->addParameter("u", u0)->addParameter("o", o2);
                     } while(0);
@@ -52,10 +52,10 @@ namespace disasm {
                     do {
                         rv = new scalar32_insn(string("addcmpb")+cc,
                                                "r{d}, {i}, {u}, $+({o}*2)");
-                        vc4_parameter d3(ParameterTypes::REGISTER, (insn >> 16) & 0x0000000f);
-                        vc4_parameter i1(ParameterTypes::REGISTER, (insn >> 16) & 0x000000f0);
-                        vc4_parameter u1(ParameterTypes::REGISTER, (insn >> 8) & 0x0000003f);
-                        vc4_parameter o3(ParameterTypes::IMMEDIATE, insn & 0x000000ff);
+                        vc4_parameter d3(ParameterTypes::REGISTER, (uint32_t)((insn >> 16) & 0x0000000f));
+                        vc4_parameter i1(ParameterTypes::REGISTER, (uint32_t)((insn >> 16) & 0x000000f0));
+                        vc4_parameter u1(ParameterTypes::REGISTER, (uint32_t)((insn >> 8) & 0x0000003f));
+                        vc4_parameter o3(ParameterTypes::IMMEDIATE, (uint32_t)(insn & 0x000000ff));
                         rv->addParameter("d", d3)->addParameter("i", i1)
                             ->addParameter("u", u1)->addParameter("o", o3);
                     } while(0);
@@ -69,13 +69,13 @@ namespace disasm {
             
             if (insn & 0x00800000) {
                 rv = new scalar32_insn("bl", "$+({o}*2)");
-                vc4_parameter o(ParameterTypes::IMMEDIATE, insn & 0x0fffffff);
+                vc4_parameter o(ParameterTypes::IMMEDIATE, (uint32_t)(insn & 0x0fffffff));
                 rv->addParameter("o", o);
             } else {
                 uint8_t cc_raw = insn & 0x0F000000 >> 24;
                 string cc(condition_codes[cc_raw]);
                 rv = new scalar32_insn( string("b")+cc, "$+({o}*2)" );
-                vc4_parameter o(ParameterTypes::IMMEDIATE, insn & 0x007fffff);
+                vc4_parameter o(ParameterTypes::IMMEDIATE, (uint32_t)(insn & 0x007fffff));
                 rv->addParameter("o", o);
             }
 
@@ -88,10 +88,10 @@ namespace disasm {
             uint8_t cc_raw = (uint8_t)((insn >> 7) & 0x0000000f);
             string cc(condition_codes[cc_raw]);
             string op = ((insn >> 21) & 1)?"st":"ld";
-            vc4_parameter b(ParameterTypes::REGISTER, insn & 0x0000001f);
-            vc4_parameter a(ParameterTypes::REGISTER, (insn >> 11) & 0x0000001f);
-            vc4_parameter d(ParameterTypes::REGISTER, (insn >> 16) & 0x0000001f);
-            vc4_parameter ww(ParameterTypes::IMMEDIATE, sz);
+            vc4_parameter b(ParameterTypes::REGISTER, (uint32_t)(insn & 0x0000001f));
+            vc4_parameter a(ParameterTypes::REGISTER, (uint32_t)((insn >> 11) & 0x0000001f));
+            vc4_parameter d(ParameterTypes::REGISTER, (uint32_t)((insn >> 16) & 0x0000001f));
+            vc4_parameter ww(ParameterTypes::IMMEDIATE, (uint32_t)(sz));
             scalar32_insn *rv = new scalar32_insn(op+w+"."+cc,
                                                   "r{d}, (r{a} + r{b} << {w})");
             rv->addParameter("d", d)->addParameter("a", a)
@@ -104,9 +104,9 @@ namespace disasm {
             string w(mem_op_widths[sz]);
             string op = ((insn >> 21) & 1)?"st":"ld";
             scalar32_insn *rv = new scalar32_insn(op+w, "r{d}, (r{a}+{o})");
-            vc4_parameter o(ParameterTypes::IMMEDIATE, insn & 0x000003ff);
-            vc4_parameter a(ParameterTypes::REGISTER, (insn >> 11) & 0x0000001f);
-            vc4_parameter d(ParameterTypes::REGISTER, (insn >> 16) & 0x0000001f);
+            vc4_parameter o(ParameterTypes::IMMEDIATE, (uint32_t)(insn & 0x000003ff));
+            vc4_parameter a(ParameterTypes::REGISTER, (uint32_t)((insn >> 11) & 0x0000001f));
+            vc4_parameter d(ParameterTypes::REGISTER, (uint32_t)((insn >> 16) & 0x0000001f));
             rv->addParameter("d", d)->addParameter("a", a)
                 ->addParameter("o", o);
             return rv;
@@ -132,8 +132,8 @@ namespace disasm {
                     break;
             }
 
-            vc4_parameter d(ParameterTypes::REGISTER, (insn >> 16) & 0x0000001f);
-            vc4_parameter o(ParameterTypes::IMMEDIATE, insn & 0x0000ffff);
+            vc4_parameter d(ParameterTypes::REGISTER, (uint32_t)((insn >> 16) & 0x0000001f));
+            vc4_parameter o(ParameterTypes::IMMEDIATE, (uint32_t)(insn & 0x0000ffff));
             scalar32_insn *rv = new scalar32_insn(op+w, fmt);
             rv->addParameter("d", d)->addParameter("o", o);
             return rv;
@@ -154,8 +154,8 @@ namespace disasm {
             string opcode(op+w+"."+cc);
             string format("r{d}, ");
             format += fd;
-            vc4_parameter a(ParameterTypes::REGISTER, (insn >> 11) & 0x0000001f);
-            vc4_parameter d(ParameterTypes::REGISTER, (insn >> 16) & 0x0000001f);
+            vc4_parameter a(ParameterTypes::REGISTER, (uint32_t)((insn >> 11) & 0x0000001f));
+            vc4_parameter d(ParameterTypes::REGISTER, (uint32_t)((insn >> 16) & 0x0000001f));
             scalar32_insn *rv = new scalar32_insn(opcode, format);
             rv->addParameter("d", d)->addParameter("a", a);
             return rv;
@@ -175,9 +175,9 @@ namespace disasm {
                 fmt = "r{d}, {o}";
             }
 
-            vc4_parameter o(ParameterTypes::IMMEDIATE, insn & 0x0000ffff);
-            vc4_parameter d(ParameterTypes::REGISTER, (insn >> 16) & 0x0000001f);
-            vc4_parameter s(ParameterTypes::REGISTER, (insn >> 21) & 0x0000001f);
+            vc4_parameter o(ParameterTypes::IMMEDIATE, (uint32_t)(insn & 0x0000ffff));
+            vc4_parameter d(ParameterTypes::REGISTER, (uint32_t)((insn >> 16) & 0x0000001f));
+            vc4_parameter s(ParameterTypes::REGISTER, (uint32_t)((insn >> 21) & 0x0000001f));
             scalar32_insn *rv = new scalar32_insn(op, fmt);
             rv->addParameter("o", o)->addParameter("d", d)
                 ->addParameter("s", s);
@@ -207,8 +207,8 @@ namespace disasm {
             uint8_t cc_raw = (uint8_t)((insn >> 7) & 0x0000000f);
             string cc(condition_codes[cc_raw]);
 
-            vc4_parameter a(ParameterTypes::REGISTER, (insn >> 11) & 0x0000001f );
-            vc4_parameter d(ParameterTypes::REGISTER, (insn >> 16) & 0x0000001f );
+            vc4_parameter a(ParameterTypes::REGISTER, (uint32_t)((insn >> 11) & 0x0000001f ));
+            vc4_parameter d(ParameterTypes::REGISTER, (uint32_t)((insn >> 16) & 0x0000001f ));
             vc4_parameter pr(pt, param);
             
             scalar32_insn *rv = new scalar32_insn(al_ops[operation], fmt);
@@ -287,8 +287,8 @@ namespace disasm {
 
         scalar32_insn *controlReg(uint32_t insn) {
             string fmt = ((insn >> 21) & 1)?"r{d}, p{a}":"p{d}, r{a}";
-            vc4_parameter a(ParameterTypes::REGISTER, insn & 0x0000001f);
-            vc4_parameter d(ParameterTypes::REGISTER, (insn >> 16) & 0x0000001f);
+            vc4_parameter a(ParameterTypes::REGISTER, (uint32_t)(insn & 0x0000001f));
+            vc4_parameter d(ParameterTypes::REGISTER, (uint32_t)((insn >> 16) & 0x0000001f));
             scalar32_insn *rv = new scalar32_insn("mov", fmt);
             rv->addParameter("a", a)->addParameter("d", d);
             return rv;
